refactor(executor): further decompose update/mod.rs into focused modules
Extract functionality from update/mod.rs (1,895 lines) into focused modules:

- executor.rs: Core execute_internal() logic with validation functions
- fast_path.rs: Single-row PK update optimizations (super-fast and fast paths)
- triggers.rs: Trigger execution and view update handling
- index_sync.rs: REPLACE conflict detection and cross-update uniqueness

mod.rs is now a thin facade (~148 lines) exposing only the public API:
- UpdateExecutor struct with execute(), execute_with_*() methods
- Re-export of execute_update_with_trigger_context()
